+++ packages/dumbbell-plot/src/components/dumbbell-plot/dumbbell-plot-props.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, 2021, 2022, 2023 Visa, Inc.
+ * Copyright (c) 2020, 2021, 2022, 2023, 2024 Visa, Inc.
  *
  * This source code is licensed under the MIT license
  * https://github.com/visa/visa-chart-components/blob/master/LICENSE
@@ -27,7 +27,7 @@ export interface IDumbbellPlotProps {
   mainTitle: string;
 
   /**
-   * @shortDescription Subtitle of the chart. A string or an array of objects. Objects should contain text (one/multiple word(s)/number(s) as a single string) and color (HEX) keys. Optional key: index.
+   * @shortDescription Subtitle of the chart. A string or an array of objects. Objects should contain text (one/multiple word(s)/number(s) as a single string) and color (HEX) keys. Optional key: index and mode (text | background).
    * @controlName TextField
    * @groupName Base */
   subTitle: string | ISubTitleType;
@@ -341,8 +341,9 @@ export interface IDumbbellPlotProps {
 export interface IDataLabelType {
   visible: boolean;
   placement?: string;
+  displayOnly?: string | string[];
   labelAccessor?: string;
-  format?: any;
+  format?: string;
   collisionHideOnly?: boolean;
   collisionPlacement?: string;
 }